adds price service and changes to the liquibase, to allow double save in mysql double precision(19) == float(19) in mysql

+++ oxm/oxm-model/src/main/java/org/agric/oxm/model/Price.java
@@ -12,67 +12,67 @@ import javax.persistence.Table;
 @Table(name = "price")
 public class Price extends BaseData {
 
-    private Crop crop;
+	private Crop crop;
 
-    private SellingPlace sellingPlace;
+	private SellingPlace sellingPlace;
 
-    private Concept sellType;
+	private Concept sellType;
 
-    private Concept unitOfMeasure;
+	private Concept unitOfMeasure;
 
-    private int price;
+	private Double price;
 
-    public Price() {
-	super();
-    }
+	public Price() {
+		super();
+	}
 
-    @OneToOne
-    @JoinColumn(name = "crop_id", nullable = false)
-    public Crop getCrop() {
-	return crop;
-    }
+	@OneToOne
+	@JoinColumn(name = "crop_id", nullable = false)
+	public Crop getCrop() {
+		return crop;
+	}
 
-    public void setCrop(Crop crop) {
-	this.crop = crop;
-    }
+	public void setCrop(Crop crop) {
+		this.crop = crop;
+	}
 
-    @ManyToOne
-    @JoinColumn(name = "selling_price_id", nullable = false)
-    public SellingPlace getSellingPlace() {
-	return sellingPlace;
-    }
+	@ManyToOne
+	@JoinColumn(name = "selling_price_id", nullable = false)
+	public SellingPlace getSellingPlace() {
+		return sellingPlace;
+	}
 
-    public void setSellingPlace(SellingPlace sellingPlace) {
-	this.sellingPlace = sellingPlace;
-    }
+	public void setSellingPlace(SellingPlace sellingPlace) {
+		this.sellingPlace = sellingPlace;
+	}
 
-    @ManyToOne(cascade = { CascadeType.PERSIST, CascadeType.MERGE })
-    @JoinColumn(name = "selling_type", nullable = true)
-    public Concept getSellType() {
-	return sellType;
-    }
+	@ManyToOne(cascade = { CascadeType.PERSIST, CascadeType.MERGE })
+	@JoinColumn(name = "selling_type", nullable = true)
+	public Concept getSellType() {
+		return sellType;
+	}
 
-    public void setSellType(Concept sellType) {
-	this.sellType = sellType;
-    }
+	public void setSellType(Concept sellType) {
+		this.sellType = sellType;
+	}
 
-    @ManyToOne(cascade = { CascadeType.PERSIST, CascadeType.MERGE })
-    @JoinColumn(name = "unit_of_measure", nullable = true)
-    public Concept getUnitOfMeasure() {
-	return unitOfMeasure;
-    }
+	@ManyToOne(cascade = { CascadeType.PERSIST, CascadeType.MERGE })
+	@JoinColumn(name = "unit_of_measure", nullable = true)
+	public Concept getUnitOfMeasure() {
+		return unitOfMeasure;
+	}
 
-    public void setUnitOfMeasure(Concept unitOfMeasure) {
-	this.unitOfMeasure = unitOfMeasure;
-    }
+	public void setUnitOfMeasure(Concept unitOfMeasure) {
+		this.unitOfMeasure = unitOfMeasure;
+	}
 
-    @Column(name = "price", nullable = true)
-    public int getPrice() {
-	return price;
-    }
+	@Column(name = "price", nullable = true)
+	public Double getPrice() {
+		return price;
+	}
 
-    public void setPrice(int price) {
-	this.price = price;
-    }
+	public void setPrice(Double price) {
+		this.price = price;
+	}
 
 }
